Currently, XYZ2npgt is unused.
Its a file that should take in an xyz file and then output a ground truth image with specified weights and shapes for atomic markers. Due to CLTEM weirdness, you need both a structural file as well as a json in order to achieve the correct padding. I haven't implemented that yet though.